Refuse the CD-Extra session gap when it does not fit, instead of publishing garbage
A data track in LAST position is read as a CD-Extra second session, and
libcdio reports the inter-session link area as part of the preceding audio
track -- so cyanrip takes 11400 frames off that track's end. The subtraction
was unguarded.

On a TOC where the gap does not fit, the LSN went negative, discid.c then
left-shifted a negative int (undefined behaviour; UBSan names discid.c:87),
and the run published

    toc=1+2+4294956496+150+375        CDDB ID: FFFF6E02

at exit 0 with no diagnostic of any kind in a default build. Nothing else
looked wrong -- three tracks, the correct total time -- so this is a confident
wrong field in an archival record, which is the one outcome this program is
built not to produce.

Measured, not reasoned about: tests/fixtures/ecd.cue is the image that does
it, and it is the first fixture with a data track in last position. mixed.cue
puts it first, which is a different layout and never reaches this branch.

Found while answering a NEXT-ROUND question from Platterpus about two-session
TOCs. They asked because a mishandled session gap shifts every sector number
and silently breaks the disc ID, and with it AccurateRip and CTDB, across a
whole class of discs. The question was worth asking and the answer turned out
to be worse than "we handle it".

Now the gap is applied only when the preceding audio track can contain it, and
when it cannot the run says so at column 0 and leaves the TOC alone. The
leadout is then a real number rather than 2^32 minus one.

The `i > 0` term is defence rather than a fix: a data-only disc never reaches
this code because cdio_cddap_open() refuses it first ("Unable to open
device!", exit 1). Measured, and written into the comment because nothing else
asserts that ordering.

discid.c's shift is now done in uint32_t. That changes no observable
behaviour and carries no revert-proof, because the guard above makes a
negative operand unreachable -- it is there because the line is two hops from
a TOC-derived value and a left shift of a negative int is undefined behaviour
rather than a wrong number.

Revert-proved by neutralising the guard alone: the edit was confirmed landed,
the build was confirmed green, and three independent checks fail -- the
missing diagnostic, the leadout/End-LSN invariant (4294956496 against a track
ending at 449), and the CDDB length field (65390 against 6). The invariant is
the one that matters: it compares two fields of a single artifact, so a
plausible-looking wrong number cannot satisfy it.

Not fixed here, because it is log text and log text is contract surface: when
the adjustment DOES apply, the per-track line reports it as
`End LSN: 449 (with offset: -10951)`, reusing the read-offset label for an
11400-frame session adjustment. One label, two causes. It goes to the round
rather than into a drive-by reword.

PROVIDER-CONTRACT.md is regenerated here because the new refusal is a
`cyanrip_log()` call and therefore P5 surface. Its banner names the build that
produced it, which is not this commit and cannot be -- no generated artifact
can carry the hash of a tree containing itself. The anchor that DOES hold is
the source hash it prints, `sha256/16 = 6a0a698965c8a112`, which recomputes
over this commit's own src/ and is what any file:line in it resolves against.

+++ src/discid.c
@@ -84,7 +84,13 @@ int crip_fill_discid(cyanrip_ctx *ctx)
     }
 
     cddb  = (cddb % 0xff) << 24;
-    cddb |= (last/75 - (ctx->tracks[0].start_lsn + 150)/75) << 8;
+    /* Shifted as unsigned. `last` cannot be negative once the CD-Extra gap is
+     * guarded in cyanrip_main.c, so this changes no observable behaviour and
+     * carries no revert-proof -- it is here because a left shift of a negative
+     * int is undefined behaviour rather than a wrong number, and this line is
+     * two hops from a value derived from the TOC. Reverting the guard alone
+     * reproduces UBSan's diagnostic at exactly this line. */
+    cddb |= ((uint32_t)(last/75 - (ctx->tracks[0].start_lsn + 150)/75)) << 8;
     cddb |= ctx->tracks[last_audio_track_idx].number;
 
     snprintf(temp, sizeof(temp), "%08X", cddb);
